+++ packages/account-tree-controller/src/AccountTreeController.ts
@@ -15,8 +15,8 @@ import type {
   MultichainAccountGroup,
 } from '@metamask/account-api';
 import { type AccountId } from '@metamask/accounts-controller';
-import type { StateMetadata } from '@metamask/base-controller';
-import { BaseController } from '@metamask/base-controller';
+import type { StateMetadata } from '@metamask/base-controller/next';
+import { BaseController } from '@metamask/base-controller/next';
 import type { TraceCallback } from '@metamask/controller-utils';
 import type { KeyringAccount } from '@metamask/keyring-api';
 import { isEvmAccountType } from '@metamask/keyring-api';
@@ -62,31 +62,31 @@ const accountTreeControllerMetadata: StateMetadata<AccountTreeControllerState> =
     accountTree: {
       includeInStateLogs: true,
       persist: false, // We do re-recompute this state everytime.
-      anonymous: false,
+      includeInDebugSnapshot: false,
       usedInUi: true,
     },
     isAccountTreeSyncingInProgress: {
       includeInStateLogs: false,
       persist: false,
-      anonymous: false,
+      includeInDebugSnapshot: false,
       usedInUi: true,
     },
     hasAccountTreeSyncingSyncedAtLeastOnce: {
       includeInStateLogs: true,
       persist: true,
-      anonymous: false,
+      includeInDebugSnapshot: false,
       usedInUi: true,
     },
     accountGroupsMetadata: {
       includeInStateLogs: true,
       persist: true,
-      anonymous: false,
+      includeInDebugSnapshot: false,
       usedInUi: true,
     },
     accountWalletsMetadata: {
       includeInStateLogs: true,
       persist: true,
-      anonymous: false,
+      includeInDebugSnapshot: false,
       usedInUi: true,
     },
   };
@@ -149,6 +149,13 @@ export class AccountTreeController extends BaseController<
 
   readonly #backupAndSyncConfig: AccountTreeControllerInternalBackupAndSyncConfig;
 
+  /**
+   * Callbacks to migrate hidden and pinned account information from the account order controller
+   */
+  readonly #accountOrderCallbacks:
+    | AccountTreeControllerConfig['accountOrderCallbacks']
+    | undefined;
+
   #initialized: boolean;
 
   /**
@@ -191,11 +198,11 @@ export class AccountTreeController extends BaseController<
     // Rules to apply to construct the wallets tree.
     this.#rules = [
       // 1. We group by entropy-source
-      new EntropyRule(this.messagingSystem),
+      new EntropyRule(this.messenger),
       // 2. We group by Snap ID
-      new SnapRule(this.messagingSystem),
+      new SnapRule(this.messenger),
       // 3. We group by wallet type (this rule cannot fail and will group all non-matching accounts)
-      new KeyringRule(this.messagingSystem),
+      new KeyringRule(this.messenger),
     ];
 
     // Initialize trace function
@@ -211,33 +218,33 @@ export class AccountTreeController extends BaseController<
       },
     };
 
+    // Used when migrating initial hidden/pinned state for groups (if available).
+    this.#accountOrderCallbacks = config?.accountOrderCallbacks;
+
     // Initialize the backup and sync service
     this.#backupAndSyncService = new BackupAndSyncService(
       this.#createBackupAndSyncContext(),
     );
 
-    this.messagingSystem.subscribe(
-      'AccountsController:accountAdded',
-      (account) => {
-        this.#handleAccountAdded(account);
-      },
-    );
+    this.messenger.subscribe('AccountsController:accountAdded', (account) => {
+      this.#handleAccountAdded(account);
+    });
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'AccountsController:accountRemoved',
       (accountId) => {
         this.#handleAccountRemoved(accountId);
       },
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'AccountsController:selectedAccountChange',
       (account) => {
         this.#handleSelectedAccountChange(account);
       },
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'UserStorageController:stateChange',
       (userStorageControllerState) => {
         this.#backupAndSyncService.handleUserStorageStateChange(
@@ -246,21 +253,21 @@ export class AccountTreeController extends BaseController<
       },
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'MultichainAccountService:multichainAccountGroupCreated',
       (group) => {
         this.#handleMultichainAccountGroupCreatedOrUpdated(group);
       },
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'MultichainAccountService:multichainAccountGroupUpdated',
       (group) => {
         this.#handleMultichainAccountGroupCreatedOrUpdated(group);
       },
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'MultichainAccountService:walletStatusChange',
       (walletId, status) => {
         this.#handleMultichainAccountWalletStatusChange(walletId, status);
@@ -387,7 +394,7 @@ export class AccountTreeController extends BaseController<
       log(
         `Selected (initial) group is: [${this.state.accountTree.selectedAccountGroup}]`,
       );
-      this.messagingSystem.publish(
+      this.messenger.publish(
         `${controllerName}:selectedAccountGroupChange`,
         this.state.accountTree.selectedAccountGroup,
         previousSelectedAccountGroup,
@@ -517,10 +524,7 @@ export class AccountTreeController extends BaseController<
     let proposedName = ''; // Empty means there's no computed name for this group.
 
     for (const id of group.accounts) {
-      const account = this.messagingSystem.call(
-        'AccountsController:getAccount',
-        id,
-      );
+      const account = this.messenger.call('AccountsController:getAccount', id);
       if (!account || !account.metadata.name.length) {
         continue;
       }
@@ -665,7 +669,7 @@ export class AccountTreeController extends BaseController<
   ) {
     const wallet = state.accountTree.wallets[walletId];
     const group = wallet.groups[groupId];
-    const persistedGroupMetadata = state.accountGroupsMetadata[group.id];
+    const persistedGroupMetadata = state.accountGroupsMetadata[groupId];
 
     // Apply persisted name if available (including empty strings)
     if (persistedGroupMetadata?.name !== undefined) {
@@ -708,9 +712,40 @@ export class AccountTreeController extends BaseController<
     // Apply persisted UI states
     if (persistedGroupMetadata?.pinned?.value !== undefined) {
       group.metadata.pinned = persistedGroupMetadata.pinned.value;
+    } else {
+      let isPinned = false;
+
+      if (this.#accountOrderCallbacks?.isPinnedAccount) {
+        isPinned = group.accounts.some((account) =>
+          this.#accountOrderCallbacks?.isPinnedAccount?.(account),
+        );
+      }
+      state.accountGroupsMetadata[groupId] ??= {};
+      state.accountGroupsMetadata[groupId].pinned = {
+        value: isPinned,
+        lastUpdatedAt: 0,
+      };
+      // If any accounts was previously pinned, then we consider the group to be pinned as well.
+      group.metadata.pinned = isPinned;
     }
+
     if (persistedGroupMetadata?.hidden?.value !== undefined) {
       group.metadata.hidden = persistedGroupMetadata.hidden.value;
+    } else {
+      let isHidden = false;
+
+      if (this.#accountOrderCallbacks?.isHiddenAccount) {
+        isHidden = group.accounts.some((account) =>
+          this.#accountOrderCallbacks?.isHiddenAccount?.(account),
+        );
+      }
+      state.accountGroupsMetadata[groupId] ??= {};
+      state.accountGroupsMetadata[groupId].hidden = {
+        value: isHidden,
+        lastUpdatedAt: 0,
+      };
+      // If any accounts was previously hidden, then we consider the group to be hidden as well.
+      group.metadata.hidden = isHidden;
     }
   }
 
@@ -768,10 +803,7 @@ export class AccountTreeController extends BaseController<
 
     const accounts: InternalAccount[] = [];
     for (const id of group.accounts) {
-      const account = this.messagingSystem.call(
-        'AccountsController:getAccount',
-        id,
-      );
+      const account = this.messenger.call('AccountsController:getAccount', id);
 
       // For now, we're filtering undefined account, but I believe
       // throwing would be more appropriate here.
@@ -838,7 +870,7 @@ export class AccountTreeController extends BaseController<
         }
       });
 
-      this.messagingSystem.publish(
+      this.messenger.publish(
         `${controllerName}:accountTreeChange`,
         this.state.accountTree,
       );
@@ -902,14 +934,14 @@ export class AccountTreeController extends BaseController<
             }
           });
 
-          this.messagingSystem.publish(
+          this.messenger.publish(
             `${controllerName}:accountTreeChange`,
             this.state.accountTree,
           );
 
           // Emit selectedAccountGroupChange event if the selected group changed
           if (selectedAccountGroupChanged) {
-            this.messagingSystem.publish(
+            this.messenger.publish(
               `${controllerName}:selectedAccountGroupChange`,
               this.state.accountTree.selectedAccountGroup,
               previousSelectedAccountGroup,
@@ -1176,7 +1208,7 @@ export class AccountTreeController extends BaseController<
   #getMultichainAccountWallets(): MultichainAccountWallet<
     Bip44Account<KeyringAccount>
   >[] {
-    return this.messagingSystem.call(
+    return this.messenger.call(
       'MultichainAccountService:getMultichainAccountWallets',
     );
   }
@@ -1187,9 +1219,7 @@ export class AccountTreeController extends BaseController<
    * @returns The list of all internal accounts.
    */
   #listAccounts(): InternalAccount[] {
-    return this.messagingSystem.call(
-      'AccountsController:listMultichainAccounts',
-    );
+    return this.messenger.call('AccountsController:listMultichainAccounts');
   }
 
   /**
@@ -1269,7 +1299,7 @@ export class AccountTreeController extends BaseController<
       `Selected group is now: [${this.state.accountTree.selectedAccountGroup}]`,
     );
 
-    this.messagingSystem.publish(
+    this.messenger.publish(
       `${controllerName}:selectedAccountGroupChange`,
       groupId,
       previousSelectedAccountGroup,
@@ -1277,7 +1307,7 @@ export class AccountTreeController extends BaseController<
 
     // Update AccountsController - this will trigger selectedAccountChange event,
     // but our handler is idempotent so it won't cause infinite loop
-    this.messagingSystem.call(
+    this.messenger.call(
       'AccountsController:setSelectedAccount',
       accountToSelect,
     );
@@ -1292,7 +1322,7 @@ export class AccountTreeController extends BaseController<
   #getDefaultSelectedAccountGroup(wallets: {
     [walletId: AccountWalletId]: AccountWalletObject;
   }): AccountGroupId | '' {
-    const selectedAccount = this.messagingSystem.call(
+    const selectedAccount = this.messenger.call(
       'AccountsController:getSelectedMultichainAccount',
     );
     if (selectedAccount && selectedAccount.id) {
@@ -1334,7 +1364,7 @@ export class AccountTreeController extends BaseController<
     this.update((state) => {
       state.accountTree.selectedAccountGroup = groupId;
     });
-    this.messagingSystem.publish(
+    this.messenger.publish(
       `${controllerName}:selectedAccountGroupChange`,
       groupId,
       previousSelectedAccountGroup,
@@ -1370,7 +1400,7 @@ export class AccountTreeController extends BaseController<
         }
       }
     });
-    this.messagingSystem.publish(
+    this.messenger.publish(
       `${controllerName}:accountTreeChange`,
       this.state.accountTree,
     );
@@ -1424,7 +1454,7 @@ export class AccountTreeController extends BaseController<
     if (group) {
       let candidate;
       for (const id of group.accounts) {
-        const account = this.messagingSystem.call(
+        const account = this.messenger.call(
           'AccountsController:getAccount',
           id,
         );
@@ -1467,7 +1497,7 @@ export class AccountTreeController extends BaseController<
         }
 
         for (const id of group.accounts) {
-          const account = this.messagingSystem.call(
+          const account = this.messenger.call(
             'AccountsController:getAccount',
             id,
           );
@@ -1710,37 +1740,37 @@ export class AccountTreeController extends BaseController<
    * Registers message handlers for the AccountTreeController.
    */
   #registerMessageHandlers(): void {
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:getSelectedAccountGroup`,
       this.getSelectedAccountGroup.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:setSelectedAccountGroup`,
       this.setSelectedAccountGroup.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:getAccountsFromSelectedAccountGroup`,
       this.getAccountsFromSelectedAccountGroup.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:setAccountWalletName`,
       this.setAccountWalletName.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:setAccountGroupName`,
       this.setAccountGroupName.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:setAccountGroupPinned`,
       this.setAccountGroupPinned.bind(this),
     );
 
-    this.messagingSystem.registerActionHandler(
+    this.messenger.registerActionHandler(
       `${controllerName}:setAccountGroupHidden`,
       this.setAccountGroupHidden.bind(this),
     );
@@ -1787,7 +1817,7 @@ export class AccountTreeController extends BaseController<
     return {
       ...this.#backupAndSyncConfig,
       controller: this,
-      messenger: this.messagingSystem,
+      messenger: this.messenger,
       controllerStateUpdateFn: this.update.bind(this),
       traceFn: this.#trace.bind(this),
       groupIdToWalletId: this.#groupIdToWalletId,